new_backup_code/etl_pipeline/requirements.txt
#pyspark>=3.5.0
pyspark==3.5.0
pyyaml>=6.0
sentence-transformers>=2.2.2
pandas>=2.0.0
pyarrow>=14.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
